python/flask/forms-examples-base-v1/requirements.txt
# requirements.txt

Flask
Flask-SQLAlchemy
Flask-WTF
Bootstrap-Flask
sqlalchemy-utils
python-dotenv